requirements.txt
# local packages
-e .
# external packages
rasa==3.4.0
python-dotenv
telebot==0.0.5
setuptools==59.6.0